refactor(utils): modify utils structure
