+++ packages/vlossom/src/utils/storybook/index.ts
@@ -1,0 +1,1 @@
+export { default as storybookMetaUtils } from './meta';
