+++ packages/vlossom/src/utils/storybook/meta.ts
@@ -1,0 +1,16 @@
+const storybookMetaUtils = {
+    getMetaArguments(componentProps: { [key: string]: any }, originalArgs: { [key: string]: any } = {}) {
+        const metaArgs: { [key: string]: any } = {};
+        Object.keys(componentProps).forEach((prop) => {
+            const { default: defaultValue } = componentProps[prop];
+            if (typeof defaultValue === 'function') {
+                metaArgs[prop] = defaultValue();
+                return;
+            }
+            metaArgs[prop] = defaultValue ?? null;
+        });
+        return { ...metaArgs, ...originalArgs };
+    },
+};
+
+export default storybookMetaUtils;
