+++ packages/vlossom/src/utils/string.ts
@@ -1,0 +1,10 @@
+const stringUtil = {
+    pascalToKebab(str: string) {
+        return str
+            .split(/(?=[A-Z])/)
+            .map((s) => s.toLowerCase())
+            .join('-');
+    },
+};
+
+export default stringUtil;
